# unvs/docs-delete/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\"JWT Authorization header using the Bearer scheme. Enter your token in the format 'Bearer \u003ctoken\u003e'\"",
        "title": "Go API Example",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/accounts/create": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Tạo một tài khoản người dùng mới với username, email và mật khẩu.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Tạo một tài khoản người dùng mới",
                "parameters": [
                    {
                        "description": "Thông tin tài khoản cần tạo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tạo tài khoản thành công",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.CreateAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ (validation errors)",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email đã tồn tại",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi nội bộ server",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/login": {
            "post": {
                "description": "Xác thực thông tin đăng nhập của người dùng (email hoặc username) và trả về JWT token nếu thành công.",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Đăng nhập người dùng và nhận JWT token (JSON/Form)",
                "parameters": [
                    {
                        "description": "Thông tin đăng nhập (email/username và mật khẩu)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công, trả về JWT token",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ (validation errors)",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Thông tin đăng nhập không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi nội bộ server",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hz": {
            "get": {
                "description": "Trả về chuỗi \"Hello World!\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "heathz"
                ],
                "summary": "Lấy một lời chào đơn giản",
                "responses": {
                    "200": {
                        "description": "Hello World!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/oauth/token": {
            "post": {
                "description": "Xác thực thông tin đăng nhập từ form data (username và mật khẩu) và trả về JWT token nếu thành công.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Đăng nhập người dùng bằng Form Submit (Username/Password)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tên người dùng (không phải email)",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Mật khẩu",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kiểu cấp quyền (thường là 'password' cho OAuth2, tùy chọn)",
                        "name": "grant_type",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công, trả về JWT token",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ (validation errors)",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Thông tin đăng nhập không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi nội bộ server",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler_account.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_app_handler_account.CreateAccountRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_app_handler_account.CreateAccountResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/unvs_internal_model_auth.User"
                }
            }
        },
        "internal_app_handler_account.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_app_handler_account.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_app_handler_account.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "description": "Có thể bao gồm thêm thông tin người dùng nếu cần, nhưng không phải mật khẩu hash",
                    "allOf": [
                        {
                            "$ref": "#/definitions/unvs_internal_model_auth.User"
                        }
                    ]
                }
            }
        },
        "unvs_internal_model_auth.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "failedLoginCount": {
                    "type": "integer"
                },
                "isLocked": {
                    "type": "boolean"
                },
                "isSupperUser": {
                    "type": "boolean"
                },
                "lastFailedLoginAt": {
                    "type": "string"
                },
                "lastLoginAt": {
                    "type": "string"
                },
                "lastPasswordChangeAt": {
                    "type": "string"
                },
                "roleId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2FormLogin": {
            "description": "\"OAuth2 Password Flow (Form Submit) - Use for explicit form data submission.\"",
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token-form"
        },
        "OAuth2Password": {
            "description": "\"OAuth2 Password Flow - Enter email/username and password in the popup to get token.\"",
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token"
        }
    }
}